[update] Separate lib and test code for each chapters
