+++ chap02/lib.py
@@ -2,14 +2,6 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 
-def costFunction(X, y, theta):
-    m = len(X)
-    h = np.dot(X, theta)
-    return 1/(2*m) * np.sum((h - y)**2)
-
-def featureNormalize(X):
-    return (X - np.mean(X)) / np.std(X, ddof=1)
-
 def gradientDescent(X, y, theta, alpha, iterations=1000, intercept=True, history=False, debug=False):
     m, n = X.shape
     j_history = []
@@ -19,7 +11,6 @@ def gradientDescent(X, y, theta, alpha, iterations=1000, intercept=True, history
           x = X[:, 1:]
           h = theta[0, :] + np.dot(x, theta[1:, :])
           theta_zero = theta[0, :] - alpha * (1/m) * np.sum(h-y)
-          # theta_one = theta[1:, :] - alpha * (1/m) * np.sum((h-y) * x)
           theta_one = theta[1:, :] - alpha * (1/m) * np.dot(x.T, (h-y))
           theta = np.insert(theta_one, 0, theta_zero).reshape(-1, 1)
           if debug:
@@ -31,19 +22,18 @@ def gradientDescent(X, y, theta, alpha, iterations=1000, intercept=True, history
         else:
           h = np.dot(X, theta)
           theta = theta - alpha * (1/m) * np.dot(X.T, (h - y))
-          # theta = theta - alpha * (1/m) * np.sum((h - y)*X) Can you explain why this is uncorrect ?? 
           if debug:
             print(i)
             print(theta)
-        j = costFunction(X, y, theta)
+        J, grad = lrCostFunction(X, y, theta)
         
         if history:
           theta_history.append(theta)
-          j_history.append(j)
+          j_history.append(J)
     if history:
       return theta_history, j_history
     else:
-      return theta, j
+      return theta, J
 
 def hessian(Xtil, yhat, tol=1e-10):
     r = np.clip(yhat * (1 - yhat), tol, np.inf)
@@ -86,17 +76,6 @@ def newtonOptimize(Xtil, y, theta, iter=0, max_iter=10, tol=1e-10, diff=np.inf, 
         return theta_hist, J_hist
     else:
         return theta, J
-
-
-def plotCostSurface(theta0_hist, theta1_hist, j_hist):
-    
-    df = pd.DataFrame({'theta0': theta0_hist, 'theta1': theta1_hist, 'j':j_hist})
-    M, B = np.meshgrid(df['theta0'].values, df['theta1'].values)
-    Z = np.meshgrid(df['j'].values, df['j'].values)[0]
-    fig = plt.figure(figsize=(20, 10))
-    ax = fig.add_subplot(111, projection='3d')
-    surf = ax.plot_surface(M, B, Z, rstride=1, cstride=1, cmap="plasma")
-    fig.colorbar(surf)
 
 def sigmoid(X):
     z = np.multiply(X, -1)
