+++ requirements.txt
@@ -1,0 +1,33 @@
+appnope==0.1.2
+attrs==21.2.0
+backcall==0.2.0
+cycler==0.10.0
+decorator==5.0.9
+iniconfig==1.1.1
+ipython==7.23.1
+ipython-genutils==0.2.0
+jedi==0.18.0
+kiwisolver==1.3.2
+matplotlib==3.4.3
+matplotlib-inline==0.1.2
+numpy==1.20.3
+packaging==20.9
+pandas==1.2.4
+parso==0.8.2
+pexpect==4.8.0
+pickleshare==0.7.5
+Pillow==8.3.1
+pluggy==0.13.1
+prompt-toolkit==3.0.18
+ptyprocess==0.7.0
+py==1.10.0
+Pygments==2.9.0
+pyparsing==2.4.7
+pytest==6.2.4
+python-dateutil==2.8.1
+pytz==2021.1
+scipy==1.7.1
+six==1.16.0
+toml==0.10.2
+traitlets==5.0.5
+wcwidth==0.2.5
